Add source.initImage() and source.initVideo()
This allows you to patch image and video data into the Hydra synth
in a single line of code.

Due to [Cross-Origin Resource
Sharing](https://developer.mozilla.org/en-US/docs/Web/HTTP/CORS)
limits, the data needs to come from the Hydra server itself.
If you run the server locally, an easy way to do this is to create
a `hydra-server/public/content` directory and place or link media
files in there.

You can then load an image like this:

    s0.initImage("/content/myimage.png");

Or a video like this:

    s0.initVideo("/content/myvideo.mp4");